UBUNTU: SAUCE: sof-firmware: update the sof-firmware to stable-v1.6
BugLink: https://bugs.launchpad.net/bugs/1900809

Intel released the sof-firmware stable-v1.6, the tgl machines and
soundwire machines need this version, so upgrade it.

And I kept some files which are in the v1.4.2 but not in the v1.6,
we need the new updated sof-firmware to work with the old kernels
like ubuntu 5.4, 5.6 kernels, so I didn't remove them, instead put
them in the v1.4.2 folder and generate links for them.

The files are copied from the commit:
cbdec6963b2c2d58b0080955d3c11b96ff4c92f0 of
https://github.com/thesofproject/sof-bin.git stable-v1.6

Signed-off-by: Hui Wang <hui.wang@canonical.com>